+++ src/fundamental/sha256-fundamental.h
